jockey_helpers: Smarter track-query matching
The new algorithm, implemented by `check_similarity_weighted()`,
makes use of a confidence value which is calculated by averaging
results from fuzzy string matching algorithms in `thefuzz` and a
given ranking, which in this case comes from Spotify (the order in
which the tracks appear in the search results).

The resulting confidences, while not perfect, give better insight
into the closeness of a human query and a matching search result
than those given by the previous difflib-based algorithm.