feat: Allow re-enqueueing stale paused tracks
This is a workaround for a bug where Lavalink prematurely sends out
track finished events when resuming tracks that were paused for more
than a few seconds.

The new config key `bot.reenqueue_paused` and corresponding env variable
`BLANCO_REENQUEUE_PAUSED`, when set to true, will cause Blanco to re-
enqueue the resumed track after 60 seconds of being paused. This is disabled by
default as it could be fixed when Lavalink v4 comes out of beta.